Exit compositor when last client exits

+++ apps/live-session.c
@@ -5,10 +5,13 @@
 #include <errno.h>
 #include <sys/wait.h>
 #include <toaru/auth.h>
+#include <toaru/yutani.h>
 #include <toaru/trace.h>
 #define TRACE_APP_NAME "live-session"
 
 int main(int argc, char * argv[]) {
+	int pid;
+
 	if (getuid() != 0) {
 		return 1;
 	}
@@ -24,18 +27,25 @@ int main(int argc, char * argv[]) {
 		return 1;
 	}
 
-	int pid = 0;
+	/* Dummy session for live-session prevents compositor from killing itself
+	 * when the main session dies the first time. */
+	yutani_init();
+
 	do {
 		pid = wait(NULL);
 	} while ((pid > 0 && pid != _session_pid) || (pid == -1 && errno == EINTR));
 
 	TRACE("Live session has ended, launching graphical login.");
-	char * args[] = {"/bin/glogin",NULL};
-	execvp(args[0],args);
+	int _glogin_pid = fork();
+	if (!_glogin_pid) {
+		char * args[] = {"/bin/glogin",NULL};
+		execvp(args[0],args);
+		system("reboot");
+	}
 
-	TRACE("failed to start glogin after log out, trying to reboot instead.");
-	system("reboot");
-
+	do {
+		pid = wait(NULL);
+	} while ((pid > 0 && pid != _glogin_pid) || (pid == -1 && errno == EINTR));
 
 	return 0;
 }
